Major update serverside - proper handling of extension data when interfacing with simpleDB. Definition of a schema (though no robustness checking yet).
Minor changes in the extension (intermediate commit). Flow now includes authenticating via direct login rather than FB, but this is not implemented (boo) - we may want to do this with an icon in the browser window, or a popup from the context menu. Also contains stub function for getting the app name from the itunes store, but I haven't figured out how to do that yet, either. Messy messy.